Change unbounded cpu context for tid

+++ src/instrument/ctf/ctfapi/CTFTypes.hpp
@@ -14,6 +14,6 @@ typedef uint8_t  ctf_event_id_t;
 typedef uint16_t ctf_task_type_id_t;
 typedef uint32_t ctf_task_id_t;
 typedef uint16_t ctf_cpu_id_t;
-typedef uint32_t ctf_thread_id_t;
+typedef uint16_t ctf_thread_id_t;
 
 #endif //CTFTYPES_HPP
